Make debug subgraph inlining thread local
Summary: The amount of places this function needed to be threaded through
was bad, and would get works with script::Function being exposed to python.

This turns it into a thread local property, and updates the tests
that need it. This requires a lot less argument threading to achieve
the same result.

Test Plan: test_jit.py

gh-metadata: pytorch pytorch 19136 gh/zdevito/14/head